Start of MVP. Backend. Secure login atm.

+++ src/main/java/com/wheel/wheelPicker/repository/WheelConfigurationRepository.java
@@ -3,6 +3,6 @@ package com.wheel.wheelPicker.repository;
 import com.wheel.wheelPicker.model.WheelConfiguration;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WheelConfigurationRepository extends JpaRepository<WheelConfiguration, Integer> {
+public interface WheelConfigurationRepository extends JpaRepository<WheelConfiguration, Long> {
     // custom queries
 }
